+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [v0.3.1] - 2016-11-08
 
 - Fixed README and examples
+- First round to fix audio playback
 
 ## [v0.3.0] - 2016-11-08